판매업체 등록시 모든 상품정보를 매핑 테이블에 등록하기

+++ hli/src/main/java/com/hli/service/AdminService.java
@@ -28,7 +28,7 @@ public interface AdminService {
 	public List<ManagerVO> getManagerList(SearchVO search);
 	
 	//판매업체 관리
-	public long addSeller(SellerVO seller) throws PersistenceException;
+	public int addSeller(SellerVO seller) throws PersistenceException;
 	public int countSeller(SearchVO search);
 	public List<SellerVO> getSellerList(SearchVO search);
 }
